# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/ProcessAddDiet/{id}:
    put:
      tags:
      - rest-controller
      operationId: processAddDiet
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content: {}
  /api/ProcessFormSignUp:
    post:
      tags:
      - rest-controller
      operationId: processRegister
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: password
        in: query
        required: true
        schema:
          type: string
      - name: mail
        in: query
        required: true
        schema:
          type: string
      responses:
        "201":
          description: Created
          content: {}
  /api/ProcessFormRecipe:
    post:
      tags:
      - rest-controller
      operationId: processMenuMaker
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: time
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: difficulty
        in: query
        required: true
        schema:
          type: string
      - name: preparation
        in: query
        required: true
        schema:
          type: string
      - name: ingredients
        in: query
        required: true
        schema:
          type: string
      - name: booleanos
        in: query
        required: true
        schema:
          type: array
          items:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageRecipe
              type: object
              properties:
                imageRecipe:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Recipe'
  /api/ProcessFormRecipe/{id}/image:
    post:
      tags:
      - rest-controller
      operationId: uploadImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageRecipe
              type: object
              properties:
                imageRecipe:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/ProcessFormMenu:
    post:
      tags:
      - rest-controller
      operationId: processRecipeMaker
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: lunchMonday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: lunchTuesday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: lunchWednesday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: lunchThursday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: lunchFriday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: lunchSaturday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: lunchSunday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: dinnerMonday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: dinnerTuesday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: dinnerWednesday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: dinnerThursday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: dinnerFriday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: dinnerSaturday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: dinnerSunday
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Created
          content: {}
  /api/ProcessFormLogIn:
    post:
      tags:
      - rest-controller
      operationId: processForm
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: password
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content: {}
  /api/ProcessFormDiet:
    post:
      tags:
      - rest-controller
      operationId: processFormDiet
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      responses:
        "201":
          description: Created
          content: {}
  /api/ProcessAddRecipe:
    post:
      tags:
      - rest-controller
      operationId: processAddRecipe
      parameters:
      - name: id_Recipe
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Recipe'
  /api/ProcessActiveMenu:
    post:
      tags:
      - rest-controller
      operationId: processActiveMenu
      parameters:
      - name: id_Menu
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content: {}
...
  /getMoreRecipes:
    get:
      tags:
      - web-controller
      operationId: getMoreProducts
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageRecipe'
...
  /api/Recipe/{id}:
    get:
      tags:
      - rest-controller
      operationId: showRecipe
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Recipe'
  /api/Recipe/{id}/image:
    get:
      tags:
      - rest-controller
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/Menu/{id}:
    get:
      tags:
      - rest-controller
      operationId: getMenu_Selected
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content: {}
...
  /api/GetMoreRecipes:
    get:
      tags:
      - rest-controller
      operationId: getMoreProducts_1
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageRecipe'
...
  /api/Diet/{id}:
    get:
      tags:
      - rest-controller
      operationId: getDietPage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content: {}
...
  /api/ProcessRemoveRecipe:
    delete:
      tags:
      - rest-controller
      operationId: processRemoveRecipe
      parameters:
      - name: id_Recipe
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Recipe'
  /api/ProcessRemoveDiet/{id}:
    delete:
      tags:
      - rest-controller
      operationId: processRemoveDiet
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content: {}
  /api/ProcessDeleteRecipe:
    delete:
      tags:
      - rest-controller
      operationId: processDeleteRecipe
      parameters:
      - name: id_RecipeDelete
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Recipe'
components:
  schemas:
    Recipe:
      type: object
      properties:
        id:
          type: integer
          format: int64
        preparation:
          type: string
        name:
          type: string
        cookTime:
          type: integer
          format: int32
        difficulty:
          type: string
        uploadDate:
          type: string
          format: date-time
        ingredients:
          type: string
        creator:
          type: string
        hasPhoto:
          type: boolean
        vegetables:
          type: boolean
        protein:
          type: boolean
        hydrates:
          type: boolean
        carbohydrates:
          type: boolean
        highinfat:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageRecipe:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        pageable:
          $ref: '#/components/schemas/Pageable'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean